requirements.txt
torch
torchvision
opencv-python-headless
requests
beautifulsoup4
tqdm
easyocr
google-generativeai
